Full transition to CharacterManager, stat system rework

+++ Assets/Scripts/Interface/Bars/Bar.cs
@@ -8,16 +8,14 @@ public class Bar : MonoBehaviour
     [SerializeField] private Slider slider;
 
 
-    public void SetCurrentValue(int value)
+    public void SetCurrentValue(float value)
     {
         slider.value = value;
     }
-
-
-    public void SetMaxValue(int value)
+    
+    public void SetMaxValue(float value)
     {
         slider.maxValue = value;
     }
-
 
 }
